resolved comments, and added isomorphic fetch dep
