+++ .github/workflows/main.yml
@@ -10,7 +10,7 @@ on:
   workflow_dispatch:
 
 env:
-  NPM_VERSION: "1.0.4"
+  NPM_VERSION: "1.0.3"
 
 
 # A workflow run is made up of one or more jobs that can run sequentially or in parallel
