+++ package.json
@@ -1,7 +1,7 @@
 {
   "name": "@laserfiche/lf-repository-api-client",
   "type": "module",
-  "version": "1.0.4",
+  "version": "1.0.3",
   "description": "Laserfiche Repository API Client library for TypeScript.",
   "main": "dist/src/index.js",
   "types": "dist/src/index.d.ts",
@@ -19,19 +19,21 @@
   "author": "Laserfiche",
   "license": "MIT",
   "devDependencies": {
+    "@types/isomorphic-fetch": "^0.0.35",
     "@types/jest": "^27.4.0",
     "@types/mocha": "^9.1.0",
     "@types/node": "^17.0.12",
     "babel-jest": "^27.5.1",
     "jest": "^27.5.1",
     "jest-junit": "^13.0.0",
+    "isomorphic-fetch": "^3.0.0",
     "nswag": "^13.15.5",
     "ts-jest": "^27.1.3",
     "ts-node": "^10.4.0",
     "typescript": "^4.5.5"
   },
   "dependencies": {
-    "@laserfiche/lf-api-client-core": "^1.0.1",
+    "@laserfiche/lf-api-client-core": "^1.0.2",
     "@laserfiche/lf-js-utils": "^3.0.5"
   }
 }
