+++ src/ClientBase.ts
@@ -6,15 +6,9 @@ import {
   DomainUtils,
   AccessKey,
 } from '@laserfiche/lf-api-client-core';
-import { IAttributeClientEx, AttributesClient } from './AttributeClientEx';
-import { EntriesClientEx, IEntriesClientEx } from './EntriesClientEx.js';
-import { IFieldDefinitionsClientEx, FieldDefinitionClient } from './FieldDefinitionsClientEx';
-import { ISearchEx, SearchClientEx } from './SearchesClientEx.js';
-import { ITagDefinitionsEx, TagDefinitionsEx } from './TagDefinitionsEx.js';
-import { ITemplateDefinitionsEx, TemplateDefinitionsEx } from './TemplateDefinitions.js';
 class ClientBase {}
 export interface IRepositoryApiClient {
-  attributesClient: IAttributeClientEx;
+  attributesClient: IAttributeClient;
   auditReasonsClient: generated.IAuditReasonsClient;
   entriesClient: IEntriesClientEx;
   fieldDefinitionsClient: IFieldDefinitionsClientEx;
@@ -30,7 +24,7 @@ export interface IRepositoryApiClient {
 export class RepositoryApiClient implements IRepositoryApiClient {
   private baseUrl: string;
 
-  public attributesClient: IAttributeClientEx;
+  public attributesClient: IAttributeClient;
   public auditReasonsClient: generated.IAuditReasonsClient;
   public entriesClient: IEntriesClientEx;
   public fieldDefinitionsClient: IFieldDefinitionsClientEx;
@@ -60,7 +54,7 @@ export class RepositoryApiClient implements IRepositoryApiClient {
       fetch,
     };
     this.baseUrl = baseUrlDebug ?? '';
-    this.attributesClient = new AttributesClient(this.baseUrl, http);
+    this.attributesClient = new AttributesClientEx(this.baseUrl, http);
     this.auditReasonsClient = new generated.AuditReasonsClient(this.baseUrl, http);
     this.entriesClient = new EntriesClientEx(this.baseUrl, http);
     this.fieldDefinitionsClient = new FieldDefinitionClient(this.baseUrl, http);
@@ -109,7 +103,6 @@ export class RepositoryApiClientHttpHandler {
     if (this.defaultRequestHeaders) {
       init.headers = Object.assign({}, this.defaultRequestHeaders, init.headers);
     }
-
     let response: Response | undefined;
     while (retryCount <= maxRetries && shouldRetry) {
       const beforeSendResult = await this._httpRequestHandler.beforeFetchRequestAsync(url, init);
@@ -149,4 +142,931 @@ export class RepositoryApiClientHttpHandler {
 function isRetryable(response: Response, init: RequestInit): boolean {
   const isIdempotent = init.method != 'POST';
   return (response.status >= 500 || response.status == 408) && isIdempotent;
+}
+
+async function getNextLinkListing<T extends generated.IODataValueContextOfIListOfEntry>(
+  http: { fetch(url: RequestInfo, init?: RequestInit): Promise<Response> },
+  processListing: (response: Response) => Promise<T>,
+  nextLink: string,
+  maxPageSize?: number
+): Promise<T> {
+  if (!nextLink) {
+    throw new Error('Next Link is undefined');
+  }
+  const prefer = CreateMaxPageSizePreferHeaderPayload(maxPageSize);
+  let options_ = <RequestInit>{
+    method: 'GET',
+    headers: {
+      Prefer: prefer !== undefined && prefer !== null ? prefer : '',
+      Accept: 'application/json',
+    },
+  };
+  let processListingTwo = processListing.bind(http);
+  return http.fetch(nextLink, options_).then((_response: Response) => {
+    return processListingTwo(_response);
+  });
+}
+
+function CreateMaxPageSizePreferHeaderPayload(maxSize?: number): string | undefined {
+  //puts the max size into the prefer header of the GET request
+  if (!maxSize) {
+    return undefined;
+  } else {
+    return `maxpagesize=${maxSize}`;
+  }
+}
+
+export interface IAttributeClient extends generated.IAttributesClient {
+  getTrusteeAttributeKeyValuePairsNextLink(args: {
+    nextLink: string;
+    maxPageSize?: number;
+  }): Promise<generated.ODataValueContextOfListOfAttribute>;
+  GetTrusteeAttributeKeyValuePairsForEach(args: {
+    callback: (response: generated.ODataValueContextOfListOfAttribute) => Promise<boolean>;
+    repoId: string;
+    everyone?: boolean;
+    select?: string;
+    orderby?: string;
+    top?: number;
+    skip?: number;
+    count?: boolean;
+    maxPageSize?: number;
+  }): Promise<void>;
+}
+
+export class AttributesClientEx extends generated.AttributesClient implements IAttributeClient {
+  async GetTrusteeAttributeKeyValuePairsForEach(args: {
+    callback: (response: generated.ODataValueContextOfListOfAttribute) => Promise<boolean>;
+    repoId: string;
+    everyone?: boolean;
+    select?: string;
+    orderby?: string;
+    top?: number;
+    skip?: number;
+    count?: boolean;
+    maxPageSize?: number;
+  }): Promise<void> {
+    let { callback, repoId, everyone, select, orderby, top, skip, count, maxPageSize } = args;
+    var response = await this.getTrusteeAttributeKeyValuePairs({
+      repoId,
+      everyone,
+      prefer: CreateMaxPageSizePreferHeaderPayload(maxPageSize),
+      select,
+      orderby,
+      top,
+      skip,
+      count,
+    });
+    let nextLink = response.odataNextLink;
+    while ((await callback(response)) && nextLink != null) {
+      response = await getNextLinkListing<generated.ODataValueContextOfListOfAttribute>(
+        // @ts-ignore: allow sub class to use private variable from the super class
+        this.http,
+        this.processGetTrusteeAttributeKeyValuePairs,
+        nextLink,
+        maxPageSize
+      );
+      nextLink = response.odataNextLink;
+    }
+  }
+  async getTrusteeAttributeKeyValuePairsNextLink(args: {
+    nextLink: string;
+    maxPageSize?: number;
+  }): Promise<generated.ODataValueContextOfListOfAttribute> {
+    let { nextLink, maxPageSize } = args;
+    return await getNextLinkListing<generated.ODataValueContextOfListOfAttribute>(
+      // @ts-ignore: allow sub class to use private variable from the super class
+      this.http,
+      this.processGetTrusteeAttributeKeyValuePairs,
+      nextLink,
+      maxPageSize
+    );
+  }
+}
+
+
+export interface IEntriesClientEx extends generated.IEntriesClient {
+  GetEntryListingForEach(args: {
+    callback: (response: generated.ODataValueContextOfIListOfEntry) => Promise<boolean>;
+    repoId: string;
+    entryId: number;
+    groupByEntryType?: boolean;
+    fields?: string[];
+    formatFields?: boolean;
+    prefer?: string;
+    culture?: string;
+    select?: string;
+    orderby?: string;
+    top?: number;
+    skip?: number;
+    count?: boolean;
+    maxPageSize?: number;
+  }): Promise<void>;
+
+  GetFieldValuesForEach(args: {
+    callback: (response: generated.ODataValueContextOfIListOfFieldValue) => Promise<boolean>;
+    repoId: string;
+    entryId: number;
+    prefer?: string;
+    formatValue?: boolean;
+    culture?: string;
+    select?: string;
+    orderby?: string;
+    top?: number;
+    skip?: number;
+    count?: boolean;
+    maxPageSize?: number;
+  }): Promise<void>;
+
+  GetLinkValuesFromEntryForEach(args: {
+    callback: (response: generated.ODataValueContextOfIListOfWEntryLinkInfo) => Promise<boolean>;
+    repoId: string;
+    entryId: number;
+    prefer?: string;
+    select?: string;
+    orderby?: string;
+    top?: number;
+    skip?: number;
+    count?: boolean;
+    maxPageSize?: number;
+  }): Promise<void>;
+
+  GetTagsAssignedToEntryForEach(args: {
+    callback: (response: generated.ODataValueContextOfIListOfWTagInfo) => Promise<boolean>;
+    repoId: string;
+    entryId: number;
+    prefer?: string;
+    select?: string;
+    orderby?: string;
+    top?: number;
+    skip?: number;
+    count?: boolean;
+    maxPageSize?: number;
+  }): Promise<void>;
+
+  getEntryListingNextLink(args: {
+    nextLink: string;
+    maxPageSize?: number;
+  }): Promise<generated.ODataValueContextOfIListOfEntry>;
+  getFieldValuesNextLink(args: {
+    nextLink: string;
+    maxPageSize?: number;
+  }): Promise<generated.ODataValueContextOfIListOfFieldValue>;
+  getLinkValuesFromEntryNextLink(args: {
+    nextLink: string;
+    maxPageSize?: number;
+  }): Promise<generated.ODataValueContextOfIListOfWEntryLinkInfo>;
+  getTagsAssignedToEntryNextLink(args: {
+    nextLink: string;
+    maxPageSize?: number;
+  }): Promise<generated.ODataValueContextOfIListOfWTagInfo>;
+}
+export class EntriesClientEx extends generated.EntriesClient implements IEntriesClientEx {
+  async GetEntryListingForEach(args: {
+    callback: (response: generated.ODataValueContextOfIListOfEntry) => Promise<boolean>;
+    repoId: string;
+    entryId: number;
+    groupByEntryType?: boolean;
+    fields?: string[];
+    formatFields?: boolean;
+    prefer?: string;
+    culture?: string;
+    select?: string;
+    orderby?: string;
+    top?: number;
+    skip?: number;
+    count?: boolean;
+    maxPageSize?: number;
+  }): Promise<void> {
+    let {
+      callback,
+      repoId,
+      entryId,
+      groupByEntryType,
+      fields,
+      formatFields,
+      prefer,
+      culture,
+      select,
+      orderby,
+      top,
+      skip,
+      count,
+      maxPageSize,
+    } = args;
+    var response = await this.getEntryListing({
+      repoId,
+      entryId,
+      groupByEntryType,
+      fields,
+      formatFields,
+      prefer: CreateMaxPageSizePreferHeaderPayload(maxPageSize),
+      culture,
+      select,
+      orderby,
+      top,
+      skip,
+      count,
+    });
+    let nextLink = response.odataNextLink;
+    while ((await callback(response)) && nextLink != null) {
+      response = await getNextLinkListing<generated.ODataValueContextOfIListOfEntry>(
+        // @ts-ignore: allow sub class to use private variable from the super class
+        this.http,
+        this.processGetEntryListing,
+        nextLink,
+        maxPageSize
+      );
+      nextLink = response.odataNextLink;
+    }
+  }
+
+  async GetFieldValuesForEach(args: {
+    callback: (response: generated.ODataValueContextOfIListOfFieldValue) => Promise<boolean>;
+    repoId: string;
+    entryId: number;
+    prefer?: string;
+    formatValue?: boolean;
+    culture?: string;
+    select?: string;
+    orderby?: string;
+    top?: number;
+    skip?: number;
+    count?: boolean;
+    maxPageSize?: number;
+  }): Promise<void> {
+    let { callback, repoId, entryId, prefer, formatValue, culture, select, orderby, top, skip, count, maxPageSize } =
+      args;
+    var response = await this.getFieldValues({
+      repoId,
+      entryId,
+      prefer: CreateMaxPageSizePreferHeaderPayload(maxPageSize),
+      formatValue,
+      culture,
+      select,
+      orderby,
+      top,
+      skip,
+      count,
+    });
+    let nextLink = response.odataNextLink;
+    while ((await callback(response)) && nextLink != null) {
+      response = await getNextLinkListing<generated.ODataValueContextOfIListOfFieldValue>(
+        // @ts-ignore: allow sub class to use private variable from the super class
+        this.http,
+        this.processGetFieldValues,
+        nextLink,
+        maxPageSize
+      );
+      nextLink = response.odataNextLink;
+    }
+  }
+
+  async GetLinkValuesFromEntryForEach(args: {
+    callback: (response: generated.ODataValueContextOfIListOfWEntryLinkInfo) => Promise<boolean>;
+    repoId: string;
+    entryId: number;
+    prefer?: string;
+    select?: string;
+    orderby?: string;
+    top?: number;
+    skip?: number;
+    count?: boolean;
+    maxPageSize?: number;
+  }): Promise<void> {
+    let { callback, repoId, entryId, prefer, select, orderby, top, skip, count, maxPageSize } = args;
+    var response = await this.getLinkValuesFromEntry({
+      repoId,
+      entryId,
+      prefer: CreateMaxPageSizePreferHeaderPayload(maxPageSize),
+      select,
+      orderby,
+      top,
+      skip,
+      count,
+    });
+    let nextLink = response.odataNextLink;
+    while ((await callback(response)) && nextLink != null) {
+      response = await getNextLinkListing<generated.ODataValueContextOfIListOfWEntryLinkInfo>(
+        // @ts-ignore: allow sub class to use private variable from the super class
+        this.http,
+        this.processGetLinkValuesFromEntry,
+        nextLink,
+        maxPageSize
+      );
+      nextLink = response.odataNextLink;
+    }
+  }
+
+  async GetTagsAssignedToEntryForEach(args: {
+    callback: (response: generated.ODataValueContextOfIListOfWTagInfo) => Promise<boolean>;
+    repoId: string;
+    entryId: number;
+    prefer?: string;
+    select?: string;
+    orderby?: string;
+    top?: number;
+    skip?: number;
+    count?: boolean;
+    maxPageSize?: number;
+  }): Promise<void> {
+    let { callback, repoId, entryId, prefer, select, orderby, top, skip, count, maxPageSize } = args;
+    var response = await this.getTagsAssignedToEntry({
+      repoId,
+      entryId,
+      prefer: CreateMaxPageSizePreferHeaderPayload(maxPageSize),
+      select,
+      orderby,
+      top,
+      skip,
+      count,
+    });
+    let nextLink = response.odataNextLink;
+    while ((await callback(response)) && nextLink != null) {
+      response = await getNextLinkListing<generated.ODataValueContextOfIListOfWTagInfo>(
+        // @ts-ignore: allow sub class to use private variable from the super class
+        this.http,
+        this.processGetTagsAssignedToEntry,
+        nextLink,
+        maxPageSize
+      );
+      nextLink = response.odataNextLink;
+    }
+  }
+
+  async getEntryListingNextLink(args: {
+    nextLink: string;
+    maxPageSize?: number;
+  }): Promise<generated.ODataValueContextOfIListOfEntry> {
+    let { nextLink, maxPageSize } = args;
+    return await getNextLinkListing<generated.ODataValueContextOfIListOfEntry>(
+      // @ts-ignore: allow sub class to use private variable from the super class
+      this.http,
+      this.processGetEntryListing,
+      nextLink,
+      maxPageSize
+    );
+  }
+
+  async getFieldValuesNextLink(args: {
+    nextLink: string;
+    maxPageSize?: number;
+  }): Promise<generated.ODataValueContextOfIListOfFieldValue> {
+    let { nextLink, maxPageSize } = args;
+    return await getNextLinkListing<generated.ODataValueContextOfIListOfFieldValue>(
+      // @ts-ignore: allow sub class to use private variable from the super class
+      this.http,
+      this.processGetFieldValues,
+      nextLink,
+      maxPageSize
+    );
+  }
+
+  async getLinkValuesFromEntryNextLink(args: {
+    nextLink: string;
+    maxPageSize?: number;
+  }): Promise<generated.ODataValueContextOfIListOfWEntryLinkInfo> {
+    let { nextLink, maxPageSize } = args;
+    return await getNextLinkListing<generated.ODataValueContextOfIListOfWEntryLinkInfo>(
+      // @ts-ignore: allow sub class to use private variable from the super class
+      this.http,
+      this.processGetLinkValuesFromEntry,
+      nextLink,
+      maxPageSize
+    );
+  }
+
+  async getTagsAssignedToEntryNextLink(args: {
+    nextLink: string;
+    maxPageSize?: number;
+  }): Promise<generated.ODataValueContextOfIListOfWTagInfo> {
+    let { nextLink, maxPageSize } = args;
+    return await getNextLinkListing<generated.ODataValueContextOfIListOfWEntryLinkInfo>(
+      // @ts-ignore: allow sub class to use private variable from the super class
+      this.http,
+      this.processGetTagsAssignedToEntry,
+      nextLink,
+      maxPageSize
+    );
+  }
+}
+
+
+export interface IFieldDefinitionsClientEx extends generated.IFieldDefinitionsClient {
+  GetFieldDefinitionsForEach(args: {
+    callback: (response: generated.ODataValueContextOfIListOfWFieldInfo) => Promise<boolean>;
+    repoId: string;
+    prefer?: string;
+    culture?: string;
+    select?: string;
+    orderby?: string;
+    top?: number;
+    skip?: number;
+    count?: boolean;
+    maxPageSize?: number;
+  }): Promise<void>;
+  getFieldDefinitionsNextLink(args: {
+    nextLink: string;
+    maxPageSize?: number;
+  }): Promise<generated.ODataValueContextOfIListOfEntry>;
+}
+
+export class FieldDefinitionClient extends generated.FieldDefinitionsClient implements IFieldDefinitionsClientEx {
+  async GetFieldDefinitionsForEach(args: {
+    callback: (response: generated.ODataValueContextOfIListOfWFieldInfo) => Promise<boolean>;
+    repoId: string;
+    prefer?: string;
+    culture?: string;
+    select?: string;
+    orderby?: string;
+    top?: number;
+    skip?: number;
+    count?: boolean;
+    maxPageSize?: number;
+  }): Promise<void> {
+    let { callback, repoId, prefer, culture, select, orderby, top, skip, count, maxPageSize } = args;
+    var response = await this.getFieldDefinitions({
+      repoId,
+      prefer: CreateMaxPageSizePreferHeaderPayload(maxPageSize),
+      culture,
+      select,
+      orderby,
+      top,
+      skip,
+      count,
+    });
+    let nextLink = response.odataNextLink;
+    while ((await callback(response)) && nextLink != null) {
+      response = await getNextLinkListing<generated.ODataValueContextOfIListOfWFieldInfo>(
+        // @ts-ignore: allow sub class to use private variable from the super class
+        this.http,
+        this.processGetFieldDefinitions,
+        nextLink,
+        maxPageSize
+      );
+      nextLink = response.odataNextLink;
+    }
+  }
+  async getFieldDefinitionsNextLink(args: {
+    nextLink: string;
+    maxPageSize?: number;
+  }): Promise<generated.ODataValueContextOfIListOfWFieldInfo> {
+    let { nextLink, maxPageSize } = args;
+    return await getNextLinkListing<generated.ODataValueContextOfIListOfWFieldInfo>(
+      // @ts-ignore: allow sub class to use private variable from the super class
+      this.http,
+      this.processGetFieldDefinitions,
+      nextLink,
+      maxPageSize
+    );
+  }
+}
+
+
+export interface ISearchEx extends generated.ISearchesClient {
+  GetSearchResultsForEach(args: {
+    callback: (response: generated.ODataValueContextOfIListOfEntry) => Promise<boolean>;
+    repoId: string;
+    searchToken: string;
+    groupByEntryType?: boolean;
+    refresh?: boolean;
+    fields?: string[];
+    formatFields?: boolean;
+    prefer?: string;
+    culture?: string;
+    select?: string;
+    orderby?: string;
+    top?: number;
+    skip?: number;
+    count?: boolean;
+    maxPageSize?: number;
+  }): Promise<void>;
+  GetSearchContextHitsForEach(args: {
+    callback: (response: generated.ODataValueContextOfIListOfContextHit) => Promise<boolean>;
+    repoId: string;
+    searchToken: string;
+    rowNumber: number;
+    prefer?: string;
+    select?: string;
+    orderby?: string;
+    top?: number;
+    skip?: number;
+    count?: boolean;
+    maxPageSize?: number;
+  }): Promise<void>;
+  GetSearchResultsNextLink(args: {
+    nextLink: string;
+    maxPageSize?: number;
+  }): Promise<generated.ODataValueContextOfIListOfEntry>;
+  GetSearchContextHitsNextLink(args: {
+    nextLink: string;
+    maxPageSize?: number;
+  }): Promise<generated.ODataValueContextOfIListOfContextHit>;
+}
+
+export class SearchClientEx extends generated.SearchesClient implements ISearchEx {
+  async GetSearchResultsForEach(args: {
+    callback: (response: generated.ODataValueContextOfIListOfEntry) => Promise<boolean>;
+    repoId: string;
+    searchToken: string;
+    groupByEntryType?: boolean;
+    refresh?: boolean;
+    fields?: string[];
+    formatFields?: boolean;
+    prefer?: string;
+    culture?: string;
+    select?: string;
+    orderby?: string;
+    top?: number;
+    skip?: number;
+    count?: boolean;
+    maxPageSize?: number;
+  }): Promise<void> {
+    let {
+      callback,
+      repoId,
+      searchToken,
+      groupByEntryType,
+      refresh,
+      fields,
+      formatFields,
+      prefer,
+      culture,
+      select,
+      orderby,
+      top,
+      skip,
+      count,
+      maxPageSize,
+    } = args;
+    var response = await this.getSearchResults({
+      repoId,
+      searchToken,
+      groupByEntryType,
+      refresh,
+      fields,
+      formatFields,
+      prefer: CreateMaxPageSizePreferHeaderPayload(maxPageSize),
+      culture,
+      select,
+      orderby,
+      top,
+      skip,
+      count,
+    });
+    let nextLink = response.odataNextLink;
+    while ((await callback(response)) && nextLink != null) {
+      response = await getNextLinkListing<generated.ODataValueContextOfIListOfEntry>(
+        // @ts-ignore: allow sub class to use private variable from the super class
+        this.http,
+        this.processGetSearchResults,
+        nextLink,
+        maxPageSize
+      );
+      nextLink = response.odataNextLink;
+    }
+  }
+  async GetSearchContextHitsForEach(args: {
+    callback: (response: generated.ODataValueContextOfIListOfContextHit) => Promise<boolean>;
+    repoId: string;
+    searchToken: string;
+    rowNumber: number;
+    prefer?: string;
+    select?: string;
+    orderby?: string;
+    top?: number;
+    skip?: number;
+    count?: boolean;
+    maxPageSize?: number;
+  }): Promise<void> {
+    let { callback, repoId, searchToken, rowNumber, prefer, select, orderby, top, skip, count, maxPageSize } = args;
+    var response = await this.getSearchContextHits({
+      repoId,
+      searchToken,
+      rowNumber,
+      prefer: CreateMaxPageSizePreferHeaderPayload(maxPageSize),
+      select,
+      orderby,
+      top,
+      skip,
+      count,
+    });
+    let nextLink = response.odataNextLink;
+    while ((await callback(response)) && nextLink != null) {
+      response = await getNextLinkListing<generated.ODataValueContextOfIListOfContextHit>(
+        // @ts-ignore: allow sub class to use private variable from the super class
+        this.http,
+        this.processGetSearchContextHits,
+        nextLink,
+        maxPageSize
+      );
+      nextLink = response.odataNextLink;
+    }
+  }
+  async GetSearchResultsNextLink(args: {
+    nextLink: string;
+    maxPageSize?: number;
+  }): Promise<generated.ODataValueContextOfIListOfEntry> {
+    let { nextLink, maxPageSize } = args;
+    return await getNextLinkListing<generated.ODataValueContextOfIListOfEntry>(
+      // @ts-ignore: allow sub class to use private variable from the super class
+      this.http,
+      this.processGetSearchResults,
+      nextLink,
+      maxPageSize
+    );
+  }
+  async GetSearchContextHitsNextLink(args: {
+    nextLink: string;
+    maxPageSize?: number;
+  }): Promise<generated.ODataValueContextOfIListOfContextHit> {
+    let { nextLink, maxPageSize } = args;
+    return await getNextLinkListing<generated.ODataValueContextOfIListOfContextHit>(
+      // @ts-ignore: allow sub class to use private variable from the super class
+      this.http,
+      this.processGetSearchContextHits,
+      nextLink,
+      maxPageSize
+    );
+  }
+}
+
+export interface ITagDefinitionsEx extends generated.ITagDefinitionsClient {
+  GetTagDefinitionsForEach(args: {
+    callback: (response: generated.ODataValueContextOfIListOfWTagInfo) => Promise<boolean>;
+    repoId: string;
+    prefer?: string;
+    culture?: string;
+    select?: string;
+    orderby?: string;
+    top?: number;
+    skip?: number;
+    count?: boolean;
+    maxPageSize?: number;
+  }): Promise<void>;
+  getTagDefinitionsNextLink(args: {
+    nextLink: string;
+    maxPageSize?: number;
+  }): Promise<generated.ODataValueContextOfIListOfWTagInfo>;
+}
+
+export class TagDefinitionsEx extends generated.TagDefinitionsClient {
+  async GetTagDefinitionsForEach(args: {
+    callback: (response: generated.ODataValueContextOfIListOfWTagInfo) => Promise<boolean>;
+    repoId: string;
+    prefer?: string;
+    culture?: string;
+    select?: string;
+    orderby?: string;
+    top?: number;
+    skip?: number;
+    count?: boolean;
+    maxPageSize?: number;
+  }): Promise<void> {
+    let { callback, repoId, prefer, culture, select, orderby, top, skip, count, maxPageSize } = args;
+    var response = await this.getTagDefinitions({
+      repoId,
+      prefer: CreateMaxPageSizePreferHeaderPayload(maxPageSize),
+      culture,
+      select,
+      orderby,
+      top,
+      skip,
+      count,
+    });
+    let nextLink = response.odataNextLink;
+    while ((await callback(response)) && nextLink != null) {
+      response = await getNextLinkListing<generated.ODataValueContextOfIListOfWTagInfo>(
+        // @ts-ignore: allow sub class to use private variable from the super class
+        this.http,
+        this.processGetTagDefinitions,
+        nextLink,
+        maxPageSize
+      );
+      nextLink = response.odataNextLink;
+    }
+  }
+  async getTagDefinitionsNextLink(args: {
+    nextLink: string;
+    maxPageSize?: number;
+  }): Promise<generated.ODataValueContextOfIListOfWTagInfo> {
+    let { nextLink, maxPageSize } = args;
+    return await getNextLinkListing<generated.ODataValueContextOfIListOfWTagInfo>(
+      // @ts-ignore: allow sub class to use private variable from the super class
+      this.http,
+      this.processGetTagDefinitions,
+      nextLink,
+      maxPageSize
+    );
+  }
+}
+
+
+export interface ITemplateDefinitionsEx extends generated.ITemplateDefinitionsClient {
+  GetTemplateDefinitionsForEach(args: {
+    callback: (response: generated.ODataValueContextOfIListOfWTemplateInfo) => Promise<boolean>;
+    repoId: string;
+    templateName?: string;
+    prefer?: string;
+    culture?: string;
+    select?: string;
+    orderby?: string;
+    top?: number;
+    skip?: number;
+    count?: boolean;
+    maxPageSize?: number;
+  }): Promise<void>;
+  GetTemplateFieldDefinitionsForEach(args: {
+    callback: (response: generated.ODataValueContextOfIListOfTemplateFieldInfo) => Promise<boolean>;
+    repoId: string;
+    templateId: number;
+    prefer?: string;
+    culture?: string;
+    select?: string;
+    orderby?: string;
+    top?: number;
+    skip?: number;
+    count?: boolean;
+    maxPageSize?: number;
+  }): Promise<void>;
+  GetTemplateFieldDefinitionsByTemplateNameForEach(args: {
+    callback: (response: generated.ODataValueContextOfIListOfTemplateFieldInfo) => Promise<boolean>;
+    repoId: string;
+    templateName: string;
+    prefer?: string;
+    culture?: string;
+    select?: string;
+    orderby?: string;
+    top?: number;
+    skip?: number;
+    count?: boolean;
+    maxPageSize?: number;
+  }): Promise<void>;
+  getTemplateDefinitionsNextLink(args: {
+    nextLink: string;
+    maxPageSize?: number;
+  }): Promise<generated.ODataValueContextOfIListOfWTemplateInfo>;
+  getTemplateFieldDefinitionsNextLink(args: {
+    nextLink: string;
+    maxPageSize?: number;
+  }): Promise<generated.ODataValueContextOfIListOfTemplateFieldInfo>;
+  getTemplateFieldDefinitionsByTemplateNameNextLink(args: {
+    nextLink: string;
+    maxPageSize?: number;
+  }): Promise<generated.ODataValueContextOfIListOfTemplateFieldInfo>;
+}
+
+export class TemplateDefinitionsEx extends generated.TemplateDefinitionsClient {
+  async GetTemplateDefinitionsForEach(args: {
+    callback: (response: generated.ODataValueContextOfIListOfWTemplateInfo) => Promise<boolean>;
+    repoId: string;
+    templateName?: string;
+    prefer?: string;
+    culture?: string;
+    select?: string;
+    orderby?: string;
+    top?: number;
+    skip?: number;
+    count?: boolean;
+    maxPageSize?: number;
+  }): Promise<void> {
+    let { callback, repoId, templateName, prefer, culture, select, orderby, top, skip, count, maxPageSize } = args;
+    var response = await this.getTemplateDefinitions({
+      repoId,
+      templateName,
+      prefer: CreateMaxPageSizePreferHeaderPayload(maxPageSize),
+      culture,
+      select,
+      orderby,
+      top,
+      skip,
+      count,
+    });
+    let nextLink = response.odataNextLink;
+    while ((await callback(response)) && nextLink != null) {
+      response = await getNextLinkListing<generated.ODataValueContextOfIListOfWTemplateInfo>(
+        // @ts-ignore: allow sub class to use private variable from the super class
+        this.http,
+        this.processGetTemplateDefinitions,
+        nextLink,
+        maxPageSize
+      );
+      nextLink = response.odataNextLink;
+    }
+  }
+
+  async GetTemplateFieldDefinitionsForEach(args: {
+    callback: (response: generated.ODataValueContextOfIListOfTemplateFieldInfo) => Promise<boolean>;
+    repoId: string;
+    templateId: number;
+    prefer?: string;
+    culture?: string;
+    select?: string;
+    orderby?: string;
+    top?: number;
+    skip?: number;
+    count?: boolean;
+    maxPageSize?: number;
+  }): Promise<void> {
+    let { callback, repoId, templateId, prefer, culture, select, orderby, top, skip, count, maxPageSize } = args;
+    var response = await this.getTemplateFieldDefinitions({
+      repoId,
+      templateId,
+      prefer: CreateMaxPageSizePreferHeaderPayload(maxPageSize),
+      culture,
+      select,
+      orderby,
+      top,
+      skip,
+      count,
+    });
+    let nextLink = response.odataNextLink;
+    while ((await callback(response)) && nextLink != null) {
+      response = await getNextLinkListing<generated.ODataValueContextOfIListOfTemplateFieldInfo>(
+        // @ts-ignore: allow sub class to use private variable from the super class
+        this.http,
+        this.processGetTemplateFieldDefinitions,
+        nextLink,
+        maxPageSize
+      );
+      nextLink = response.odataNextLink;
+    }
+  }
+
+  async GetTemplateFieldDefinitionsByTemplateNameForEach(args: {
+    callback: (response: generated.ODataValueContextOfIListOfTemplateFieldInfo) => Promise<boolean>;
+    repoId: string;
+    templateName: string;
+    prefer?: string;
+    culture?: string;
+    select?: string;
+    orderby?: string;
+    top?: number;
+    skip?: number;
+    count?: boolean;
+    maxPageSize?: number;
+  }): Promise<void> {
+    let { callback, repoId, templateName, prefer, culture, select, orderby, top, skip, count, maxPageSize } = args;
+    var response = await this.getTemplateFieldDefinitionsByTemplateName({
+      repoId,
+      templateName,
+      prefer: CreateMaxPageSizePreferHeaderPayload(maxPageSize),
+      culture,
+      select,
+      orderby,
+      top,
+      skip,
+      count,
+    });
+    let nextLink = response.odataNextLink;
+    while ((await callback(response)) && nextLink != null) {
+      response = await getNextLinkListing<generated.ODataValueContextOfIListOfTemplateFieldInfo>(
+        // @ts-ignore: allow sub class to use private variable from the super class
+        this.http,
+        this.processGetTemplateFieldDefinitionsByTemplateName,
+        nextLink,
+        maxPageSize
+      );
+      nextLink = response.odataNextLink;
+    }
+  }
+  async getTemplateDefinitionsNextLink(args: {
+    nextLink: string;
+    maxPageSize?: number;
+  }): Promise<generated.ODataValueContextOfIListOfWTemplateInfo> {
+    let { nextLink, maxPageSize } = args;
+    return await getNextLinkListing<generated.ODataValueContextOfIListOfWTemplateInfo>(
+      // @ts-ignore: allow sub class to use private variable from the super class
+      this.http,
+      this.processGetTemplateDefinitions,
+      nextLink,
+      maxPageSize
+    );
+  }
+  async getTemplateFieldDefinitionsNextLink(args: {
+    nextLink: string;
+    maxPageSize?: number;
+  }): Promise<generated.ODataValueContextOfIListOfTemplateFieldInfo> {
+    let { nextLink, maxPageSize } = args;
+    return await getNextLinkListing<generated.ODataValueContextOfIListOfTemplateFieldInfo>(
+      // @ts-ignore: allow sub class to use private variable from the super class
+      this.http,
+      this.processGetTemplateFieldDefinitions,
+      nextLink,
+      maxPageSize
+    );
+  }
+  async getTemplateFieldDefinitionsByTemplateNameNextLink(args: {
+    nextLink: string;
+    maxPageSize?: number;
+  }): Promise<generated.ODataValueContextOfIListOfTemplateFieldInfo> {
+    let { nextLink, maxPageSize } = args;
+    return await getNextLinkListing<generated.ODataValueContextOfIListOfTemplateFieldInfo>(
+      // @ts-ignore: allow sub class to use private variable from the super class
+      this.http,
+      this.processGetTemplateFieldDefinitionsByTemplateName,
+      nextLink,
+      maxPageSize
+    );
+  }
 }
