--- src/TemplateDefinitions.ts
@@ -1,210 +1,0 @@
-import { CreateMaxPageSizePreferHeaderPayload, getNextLinkListing } from './ClientHelper.js';
-import * as generated from './index.js';
-export interface ITemplateDefinitionsEx extends generated.ITemplateDefinitionsClient {
-  GetTemplateDefinitionsForEach(args: {
-    callback: (response: generated.ODataValueContextOfIListOfWTemplateInfo) => Promise<boolean>;
-    repoId: string;
-    templateName?: string;
-    prefer?: string;
-    culture?: string;
-    select?: string;
-    orderby?: string;
-    top?: number;
-    skip?: number;
-    count?: boolean;
-    maxPageSize?: number;
-  }): Promise<void>;
-  GetTemplateFieldDefinitionsForEach(args: {
-    callback: (response: generated.ODataValueContextOfIListOfTemplateFieldInfo) => Promise<boolean>;
-    repoId: string;
-    templateId: number;
-    prefer?: string;
-    culture?: string;
-    select?: string;
-    orderby?: string;
-    top?: number;
-    skip?: number;
-    count?: boolean;
-    maxPageSize?: number;
-  }): Promise<void>;
-  GetTemplateFieldDefinitionsByTemplateNameForEach(args: {
-    callback: (response: generated.ODataValueContextOfIListOfTemplateFieldInfo) => Promise<boolean>;
-    repoId: string;
-    templateName: string;
-    prefer?: string;
-    culture?: string;
-    select?: string;
-    orderby?: string;
-    top?: number;
-    skip?: number;
-    count?: boolean;
-    maxPageSize?: number;
-  }): Promise<void>;
-  getTemplateDefinitionsNextLink(args: {
-    nextLink: string;
-    maxPageSize?: number;
-  }): Promise<generated.ODataValueContextOfIListOfWTemplateInfo>;
-  getTemplateFieldDefinitionsNextLink(args: {
-    nextLink: string;
-    maxPageSize?: number;
-  }): Promise<generated.ODataValueContextOfIListOfTemplateFieldInfo>;
-  getTemplateFieldDefinitionsByTemplateNameNextLink(args: {
-    nextLink: string;
-    maxPageSize?: number;
-  }): Promise<generated.ODataValueContextOfIListOfTemplateFieldInfo>;
-}
-
-export class TemplateDefinitionsEx extends generated.TemplateDefinitionsClient {
-  async GetTemplateDefinitionsForEach(args: {
-    callback: (response: generated.ODataValueContextOfIListOfWTemplateInfo) => Promise<boolean>;
-    repoId: string;
-    templateName?: string;
-    prefer?: string;
-    culture?: string;
-    select?: string;
-    orderby?: string;
-    top?: number;
-    skip?: number;
-    count?: boolean;
-    maxPageSize?: number;
-  }): Promise<void> {
-    let { callback, repoId, templateName, prefer, culture, select, orderby, top, skip, count, maxPageSize } = args;
-    var response = await this.getTemplateDefinitions({
-      repoId,
-      templateName,
-      prefer: CreateMaxPageSizePreferHeaderPayload(maxPageSize),
-      culture,
-      select,
-      orderby,
-      top,
-      skip,
-      count,
-    });
-    let nextLink = response.odataNextLink;
-    while ((await callback(response)) && nextLink != null) {
-      response = await getNextLinkListing<generated.ODataValueContextOfIListOfWTemplateInfo>(
-        // @ts-ignore: allow sub class to use private variable from the super class
-        this.http,
-        this.processGetTemplateDefinitions,
-        nextLink,
-        maxPageSize
-      );
-      nextLink = response.odataNextLink;
-    }
-  }
-
-  async GetTemplateFieldDefinitionsForEach(args: {
-    callback: (response: generated.ODataValueContextOfIListOfTemplateFieldInfo) => Promise<boolean>;
-    repoId: string;
-    templateId: number;
-    prefer?: string;
-    culture?: string;
-    select?: string;
-    orderby?: string;
-    top?: number;
-    skip?: number;
-    count?: boolean;
-    maxPageSize?: number;
-  }): Promise<void> {
-    let { callback, repoId, templateId, prefer, culture, select, orderby, top, skip, count, maxPageSize } = args;
-    var response = await this.getTemplateFieldDefinitions({
-      repoId,
-      templateId,
-      prefer: CreateMaxPageSizePreferHeaderPayload(maxPageSize),
-      culture,
-      select,
-      orderby,
-      top,
-      skip,
-      count,
-    });
-    let nextLink = response.odataNextLink;
-    while ((await callback(response)) && nextLink != null) {
-      response = await getNextLinkListing<generated.ODataValueContextOfIListOfTemplateFieldInfo>(
-        // @ts-ignore: allow sub class to use private variable from the super class
-        this.http,
-        this.processGetTemplateFieldDefinitions,
-        nextLink,
-        maxPageSize
-      );
-      nextLink = response.odataNextLink;
-    }
-  }
-
-  async GetTemplateFieldDefinitionsByTemplateNameForEach(args: {
-    callback: (response: generated.ODataValueContextOfIListOfTemplateFieldInfo) => Promise<boolean>;
-    repoId: string;
-    templateName: string;
-    prefer?: string;
-    culture?: string;
-    select?: string;
-    orderby?: string;
-    top?: number;
-    skip?: number;
-    count?: boolean;
-    maxPageSize?: number;
-  }): Promise<void> {
-    let { callback, repoId, templateName, prefer, culture, select, orderby, top, skip, count, maxPageSize } = args;
-    var response = await this.getTemplateFieldDefinitionsByTemplateName({
-      repoId,
-      templateName,
-      prefer: CreateMaxPageSizePreferHeaderPayload(maxPageSize),
-      culture,
-      select,
-      orderby,
-      top,
-      skip,
-      count,
-    });
-    let nextLink = response.odataNextLink;
-    while ((await callback(response)) && nextLink != null) {
-      response = await getNextLinkListing<generated.ODataValueContextOfIListOfTemplateFieldInfo>(
-        // @ts-ignore: allow sub class to use private variable from the super class
-        this.http,
-        this.processGetTemplateFieldDefinitionsByTemplateName,
-        nextLink,
-        maxPageSize
-      );
-      nextLink = response.odataNextLink;
-    }
-  }
-  async getTemplateDefinitionsNextLink(args: {
-    nextLink: string;
-    maxPageSize?: number;
-  }): Promise<generated.ODataValueContextOfIListOfWTemplateInfo> {
-    let { nextLink, maxPageSize } = args;
-    return await getNextLinkListing<generated.ODataValueContextOfIListOfWTemplateInfo>(
-      // @ts-ignore: allow sub class to use private variable from the super class
-      this.http,
-      this.processGetTemplateDefinitions,
-      nextLink,
-      maxPageSize
-    );
-  }
-  async getTemplateFieldDefinitionsNextLink(args: {
-    nextLink: string;
-    maxPageSize?: number;
-  }): Promise<generated.ODataValueContextOfIListOfTemplateFieldInfo> {
-    let { nextLink, maxPageSize } = args;
-    return await getNextLinkListing<generated.ODataValueContextOfIListOfTemplateFieldInfo>(
-      // @ts-ignore: allow sub class to use private variable from the super class
-      this.http,
-      this.processGetTemplateFieldDefinitions,
-      nextLink,
-      maxPageSize
-    );
-  }
-  async getTemplateFieldDefinitionsByTemplateNameNextLink(args: {
-    nextLink: string;
-    maxPageSize?: number;
-  }): Promise<generated.ODataValueContextOfIListOfTemplateFieldInfo> {
-    let { nextLink, maxPageSize } = args;
-    return await getNextLinkListing<generated.ODataValueContextOfIListOfTemplateFieldInfo>(
-      // @ts-ignore: allow sub class to use private variable from the super class
-      this.http,
-      this.processGetTemplateFieldDefinitionsByTemplateName,
-      nextLink,
-      maxPageSize
-    );
-  }
-}
