+++ src/index.ts
@@ -8,6 +8,14 @@
 /* eslint-disable */
 // ReSharper disable InconsistentNaming
 
+import { UrlUtils } from '@laserfiche/lf-js-utils';
+import {
+  OAuthClientCredentialsHandler,
+  HttpRequestHandler,
+  DomainUtils,
+  AccessKey,
+} from '@laserfiche/lf-api-client-core';
+
 export interface IEntriesClient {
 
     /**
@@ -8777,4 +8785,1065 @@ function throwException(message: string, status: number, response: string, heade
         throw result;
     else
         throw new ApiException(message, status, response, headers, null);
+}
+
+class ClientBase {}
+export interface IRepositoryApiClient {
+  attributesClient: IAttributeClient;
+  auditReasonsClient: IAuditReasonsClient;
+  entriesClient: IEntriesClientEx;
+  fieldDefinitionsClient: IFieldDefinitionsClientEx;
+  repositoriesClient: IRepositoriesClient;
+  searchesClient: ISearchEx;
+  serverSessionClient: IServerSessionClient;
+  simpleSearchesClient: ISimpleSearchesClient;
+  tagDefinitionsClient: ITagDefinitionsEx;
+  tasksClient: ITasksClient;
+  templateDefinitionsClient: ITemplateDefinitionsEx;
+}
+// @ts-ignore
+export class RepositoryApiClient implements IRepositoryApiClient {
+  private baseUrl: string;
+
+  public attributesClient: IAttributeClient;
+  public auditReasonsClient: IAuditReasonsClient;
+  public entriesClient: IEntriesClientEx;
+  public fieldDefinitionsClient: IFieldDefinitionsClientEx;
+  public repositoriesClient: IRepositoriesClient;
+  public searchesClient: ISearchEx;
+  public serverSessionClient: IServerSessionClient;
+  public simpleSearchesClient: ISimpleSearchesClient;
+  public tagDefinitionsClient: ITagDefinitionsEx;
+  public tasksClient: ITasksClient;
+  public templateDefinitionsClient: ITemplateDefinitionsEx;
+
+  private repoClientHandler: RepositoryApiClientHttpHandler;
+
+  public get defaultRequestHeaders(): Record<string, string> {
+    return this.repoClientHandler.defaultRequestHeaders;
+  }
+
+  public set defaultRequestHeaders(headers: Record<string, string>) {
+    this.repoClientHandler.defaultRequestHeaders = headers;
+  }
+
+  private constructor(httpRequestHandler: HttpRequestHandler, baseUrlDebug?: string) {
+    this.repoClientHandler = new RepositoryApiClientHttpHandler(httpRequestHandler);
+    let fetch = this.repoClientHandler.httpHandler;
+    fetch = fetch.bind(this.repoClientHandler);
+    let http = {
+      fetch,
+    };
+    this.baseUrl = baseUrlDebug ?? '';
+    this.attributesClient = new AttributesClientEx(this.baseUrl, http);
+    this.auditReasonsClient = new AuditReasonsClient(this.baseUrl, http);
+    this.entriesClient = new EntriesClientEx(this.baseUrl, http);
+    this.fieldDefinitionsClient = new FieldDefinitionClient(this.baseUrl, http);
+    this.repositoriesClient = new RepositoriesClient(this.baseUrl, http);
+    this.searchesClient = new SearchClientEx(this.baseUrl, http);
+    this.serverSessionClient = new ServerSessionClient(this.baseUrl, http);
+    this.simpleSearchesClient = new SimpleSearchesClient(this.baseUrl, http);
+    this.tagDefinitionsClient = new TagDefinitionsEx(this.baseUrl, http);
+    this.tasksClient = new TasksClient(this.baseUrl, http);
+    this.templateDefinitionsClient = new TemplateDefinitionsEx(this.baseUrl, http);
+  }
+
+  public static createFromHttpRequestHandler(
+    httpRequestHandler: HttpRequestHandler,
+    baseUrlDebug?: string
+  ): RepositoryApiClient {
+    if (!httpRequestHandler) throw new Error('Argument cannot be null: httpRequestHandler');
+    let repoClient = new RepositoryApiClient(httpRequestHandler, baseUrlDebug);
+    return repoClient;
+  }
+
+  public static createFromAccessKey(
+    servicePrincipalKey: string,
+    accessKey: AccessKey,
+    baseUrlDebug?: string
+  ): RepositoryApiClient {
+    let handler = new OAuthClientCredentialsHandler(servicePrincipalKey, accessKey);
+    return RepositoryApiClient.createFromHttpRequestHandler(handler, baseUrlDebug);
+  }
+}
+/** @internal */
+export class RepositoryApiClientHttpHandler {
+  private _httpRequestHandler: HttpRequestHandler;
+  public defaultRequestHeaders: Record<string, string>;
+
+  constructor(httpRequestHandler: HttpRequestHandler) {
+    this._httpRequestHandler = httpRequestHandler;
+    this.defaultRequestHeaders = {};
+  }
+
+  public async httpHandler(url: string, init: RequestInit): Promise<Response> {
+    const maxRetries = 1;
+    let retryCount = 0;
+    let shouldRetry = true;
+
+    if (this.defaultRequestHeaders) {
+      init.headers = Object.assign({}, this.defaultRequestHeaders, init.headers);
+    }
+    let response: Response | undefined;
+    while (retryCount <= maxRetries && shouldRetry) {
+      const beforeSendResult = await this._httpRequestHandler.beforeFetchRequestAsync(url, init);
+      let absoluteUrl: string;
+      if (url.startsWith('http')) {
+        absoluteUrl = url;
+      } else {
+        const apiBasedAddress = DomainUtils.getRepositoryEndpoint(beforeSendResult.regionalDomain);
+        absoluteUrl = UrlUtils.combineURLs(apiBasedAddress, url);
+      }
+
+      try {
+        response = await fetch(absoluteUrl, init);
+        shouldRetry =
+          (await this._httpRequestHandler.afterFetchResponseAsync(absoluteUrl, response, init)) ||
+          isRetryable(response, init);
+        if (!shouldRetry) {
+          return response;
+        }
+      } catch (err) {
+        if (retryCount >= maxRetries) {
+          throw err;
+        }
+        shouldRetry = true;
+        console.warn(`Retrying fetch due to exception: ${err}`);
+      } finally {
+        retryCount++;
+      }
+    }
+    if (!response) {
+      throw new Error('Undefined response, there is a bug!');
+    }
+    return response;
+  }
+}
+
+function isRetryable(response: Response, init: RequestInit): boolean {
+  const isIdempotent = init.method != 'POST';
+  return (response.status >= 500 || response.status == 408) && isIdempotent;
+}
+
+async function getNextLinkListing<T extends IODataValueContextOfIListOfEntry>(
+  http: { fetch(url: RequestInfo, init?: RequestInit): Promise<Response> },
+  processListing: (response: Response) => Promise<T>,
+  nextLink: string,
+  maxPageSize?: number
+): Promise<T> {
+  if (!nextLink) {
+    throw new Error('Next Link is undefined');
+  }
+  const prefer = CreateMaxPageSizePreferHeaderPayload(maxPageSize);
+  let options_ = <RequestInit>{
+    method: 'GET',
+    headers: {
+      Prefer: prefer !== undefined && prefer !== null ? prefer : '',
+      Accept: 'application/json',
+    },
+  };
+  let processListingTwo = processListing.bind(http);
+  return http.fetch(nextLink, options_).then((_response: Response) => {
+    return processListingTwo(_response);
+  });
+}
+
+function CreateMaxPageSizePreferHeaderPayload(maxSize?: number): string | undefined {
+  //puts the max size into the prefer header of the GET request
+  if (!maxSize) {
+    return undefined;
+  } else {
+    return `maxpagesize=${maxSize}`;
+  }
+}
+
+export interface IAttributeClient extends IAttributesClient {
+  getTrusteeAttributeKeyValuePairsNextLink(args: {
+    nextLink: string;
+    maxPageSize?: number;
+  }): Promise<ODataValueContextOfListOfAttribute>;
+  GetTrusteeAttributeKeyValuePairsForEach(args: {
+    callback: (response: ODataValueContextOfListOfAttribute) => Promise<boolean>;
+    repoId: string;
+    everyone?: boolean;
+    select?: string;
+    orderby?: string;
+    top?: number;
+    skip?: number;
+    count?: boolean;
+    maxPageSize?: number;
+  }): Promise<void>;
+}
+
+export class AttributesClientEx extends AttributesClient implements IAttributeClient {
+  async GetTrusteeAttributeKeyValuePairsForEach(args: {
+    callback: (response: ODataValueContextOfListOfAttribute) => Promise<boolean>;
+    repoId: string;
+    everyone?: boolean;
+    select?: string;
+    orderby?: string;
+    top?: number;
+    skip?: number;
+    count?: boolean;
+    maxPageSize?: number;
+  }): Promise<void> {
+    let { callback, repoId, everyone, select, orderby, top, skip, count, maxPageSize } = args;
+    var response = await this.getTrusteeAttributeKeyValuePairs({
+      repoId,
+      everyone,
+      prefer: CreateMaxPageSizePreferHeaderPayload(maxPageSize),
+      select,
+      orderby,
+      top,
+      skip,
+      count,
+    });
+    let nextLink = response.odataNextLink;
+    while ((await callback(response)) && nextLink != null) {
+      response = await getNextLinkListing<ODataValueContextOfListOfAttribute>(
+        // @ts-ignore: allow sub class to use private variable from the super class
+        this.http,
+        this.processGetTrusteeAttributeKeyValuePairs,
+        nextLink,
+        maxPageSize
+      );
+      nextLink = response.odataNextLink;
+    }
+  }
+  async getTrusteeAttributeKeyValuePairsNextLink(args: {
+    nextLink: string;
+    maxPageSize?: number;
+  }): Promise<ODataValueContextOfListOfAttribute> {
+    let { nextLink, maxPageSize } = args;
+    return await getNextLinkListing<ODataValueContextOfListOfAttribute>(
+      // @ts-ignore: allow sub class to use private variable from the super class
+      this.http,
+      this.processGetTrusteeAttributeKeyValuePairs,
+      nextLink,
+      maxPageSize
+    );
+  }
+}
+
+export interface IEntriesClientEx extends IEntriesClient {
+  GetEntryListingForEach(args: {
+    callback: (response: ODataValueContextOfIListOfEntry) => Promise<boolean>;
+    repoId: string;
+    entryId: number;
+    groupByEntryType?: boolean;
+    fields?: string[];
+    formatFields?: boolean;
+    prefer?: string;
+    culture?: string;
+    select?: string;
+    orderby?: string;
+    top?: number;
+    skip?: number;
+    count?: boolean;
+    maxPageSize?: number;
+  }): Promise<void>;
+
+  GetFieldValuesForEach(args: {
+    callback: (response: ODataValueContextOfIListOfFieldValue) => Promise<boolean>;
+    repoId: string;
+    entryId: number;
+    prefer?: string;
+    formatValue?: boolean;
+    culture?: string;
+    select?: string;
+    orderby?: string;
+    top?: number;
+    skip?: number;
+    count?: boolean;
+    maxPageSize?: number;
+  }): Promise<void>;
+
+  GetLinkValuesFromEntryForEach(args: {
+    callback: (response: ODataValueContextOfIListOfWEntryLinkInfo) => Promise<boolean>;
+    repoId: string;
+    entryId: number;
+    prefer?: string;
+    select?: string;
+    orderby?: string;
+    top?: number;
+    skip?: number;
+    count?: boolean;
+    maxPageSize?: number;
+  }): Promise<void>;
+
+  GetTagsAssignedToEntryForEach(args: {
+    callback: (response: ODataValueContextOfIListOfWTagInfo) => Promise<boolean>;
+    repoId: string;
+    entryId: number;
+    prefer?: string;
+    select?: string;
+    orderby?: string;
+    top?: number;
+    skip?: number;
+    count?: boolean;
+    maxPageSize?: number;
+  }): Promise<void>;
+
+  getEntryListingNextLink(args: {
+    nextLink: string;
+    maxPageSize?: number;
+  }): Promise<ODataValueContextOfIListOfEntry>;
+  getFieldValuesNextLink(args: {
+    nextLink: string;
+    maxPageSize?: number;
+  }): Promise<ODataValueContextOfIListOfFieldValue>;
+  getLinkValuesFromEntryNextLink(args: {
+    nextLink: string;
+    maxPageSize?: number;
+  }): Promise<ODataValueContextOfIListOfWEntryLinkInfo>;
+  getTagsAssignedToEntryNextLink(args: {
+    nextLink: string;
+    maxPageSize?: number;
+  }): Promise<ODataValueContextOfIListOfWTagInfo>;
+}
+export class EntriesClientEx extends EntriesClient implements IEntriesClientEx {
+  async GetEntryListingForEach(args: {
+    callback: (response: ODataValueContextOfIListOfEntry) => Promise<boolean>;
+    repoId: string;
+    entryId: number;
+    groupByEntryType?: boolean;
+    fields?: string[];
+    formatFields?: boolean;
+    prefer?: string;
+    culture?: string;
+    select?: string;
+    orderby?: string;
+    top?: number;
+    skip?: number;
+    count?: boolean;
+    maxPageSize?: number;
+  }): Promise<void> {
+    let {
+      callback,
+      repoId,
+      entryId,
+      groupByEntryType,
+      fields,
+      formatFields,
+      prefer,
+      culture,
+      select,
+      orderby,
+      top,
+      skip,
+      count,
+      maxPageSize,
+    } = args;
+    var response = await this.getEntryListing({
+      repoId,
+      entryId,
+      groupByEntryType,
+      fields,
+      formatFields,
+      prefer: CreateMaxPageSizePreferHeaderPayload(maxPageSize),
+      culture,
+      select,
+      orderby,
+      top,
+      skip,
+      count,
+    });
+    let nextLink = response.odataNextLink;
+    while ((await callback(response)) && nextLink != null) {
+      response = await getNextLinkListing<ODataValueContextOfIListOfEntry>(
+        // @ts-ignore: allow sub class to use private variable from the super class
+        this.http,
+        this.processGetEntryListing,
+        nextLink,
+        maxPageSize
+      );
+      nextLink = response.odataNextLink;
+    }
+  }
+
+  async GetFieldValuesForEach(args: {
+    callback: (response: ODataValueContextOfIListOfFieldValue) => Promise<boolean>;
+    repoId: string;
+    entryId: number;
+    prefer?: string;
+    formatValue?: boolean;
+    culture?: string;
+    select?: string;
+    orderby?: string;
+    top?: number;
+    skip?: number;
+    count?: boolean;
+    maxPageSize?: number;
+  }): Promise<void> {
+    let { callback, repoId, entryId, prefer, formatValue, culture, select, orderby, top, skip, count, maxPageSize } =
+      args;
+    var response = await this.getFieldValues({
+      repoId,
+      entryId,
+      prefer: CreateMaxPageSizePreferHeaderPayload(maxPageSize),
+      formatValue,
+      culture,
+      select,
+      orderby,
+      top,
+      skip,
+      count,
+    });
+    let nextLink = response.odataNextLink;
+    while ((await callback(response)) && nextLink != null) {
+      response = await getNextLinkListing<ODataValueContextOfIListOfFieldValue>(
+        // @ts-ignore: allow sub class to use private variable from the super class
+        this.http,
+        this.processGetFieldValues,
+        nextLink,
+        maxPageSize
+      );
+      nextLink = response.odataNextLink;
+    }
+  }
+
+  async GetLinkValuesFromEntryForEach(args: {
+    callback: (response: ODataValueContextOfIListOfWEntryLinkInfo) => Promise<boolean>;
+    repoId: string;
+    entryId: number;
+    prefer?: string;
+    select?: string;
+    orderby?: string;
+    top?: number;
+    skip?: number;
+    count?: boolean;
+    maxPageSize?: number;
+  }): Promise<void> {
+    let { callback, repoId, entryId, prefer, select, orderby, top, skip, count, maxPageSize } = args;
+    var response = await this.getLinkValuesFromEntry({
+      repoId,
+      entryId,
+      prefer: CreateMaxPageSizePreferHeaderPayload(maxPageSize),
+      select,
+      orderby,
+      top,
+      skip,
+      count,
+    });
+    let nextLink = response.odataNextLink;
+    while ((await callback(response)) && nextLink != null) {
+      response = await getNextLinkListing<ODataValueContextOfIListOfWEntryLinkInfo>(
+        // @ts-ignore: allow sub class to use private variable from the super class
+        this.http,
+        this.processGetLinkValuesFromEntry,
+        nextLink,
+        maxPageSize
+      );
+      nextLink = response.odataNextLink;
+    }
+  }
+
+  async GetTagsAssignedToEntryForEach(args: {
+    callback: (response: ODataValueContextOfIListOfWTagInfo) => Promise<boolean>;
+    repoId: string;
+    entryId: number;
+    prefer?: string;
+    select?: string;
+    orderby?: string;
+    top?: number;
+    skip?: number;
+    count?: boolean;
+    maxPageSize?: number;
+  }): Promise<void> {
+    let { callback, repoId, entryId, prefer, select, orderby, top, skip, count, maxPageSize } = args;
+    var response = await this.getTagsAssignedToEntry({
+      repoId,
+      entryId,
+      prefer: CreateMaxPageSizePreferHeaderPayload(maxPageSize),
+      select,
+      orderby,
+      top,
+      skip,
+      count,
+    });
+    let nextLink = response.odataNextLink;
+    while ((await callback(response)) && nextLink != null) {
+      response = await getNextLinkListing<ODataValueContextOfIListOfWTagInfo>(
+        // @ts-ignore: allow sub class to use private variable from the super class
+        this.http,
+        this.processGetTagsAssignedToEntry,
+        nextLink,
+        maxPageSize
+      );
+      nextLink = response.odataNextLink;
+    }
+  }
+
+  async getEntryListingNextLink(args: {
+    nextLink: string;
+    maxPageSize?: number;
+  }): Promise<ODataValueContextOfIListOfEntry> {
+    let { nextLink, maxPageSize } = args;
+    return await getNextLinkListing<ODataValueContextOfIListOfEntry>(
+      // @ts-ignore: allow sub class to use private variable from the super class
+      this.http,
+      this.processGetEntryListing,
+      nextLink,
+      maxPageSize
+    );
+  }
+
+  async getFieldValuesNextLink(args: {
+    nextLink: string;
+    maxPageSize?: number;
+  }): Promise<ODataValueContextOfIListOfFieldValue> {
+    let { nextLink, maxPageSize } = args;
+    return await getNextLinkListing<ODataValueContextOfIListOfFieldValue>(
+      // @ts-ignore: allow sub class to use private variable from the super class
+      this.http,
+      this.processGetFieldValues,
+      nextLink,
+      maxPageSize
+    );
+  }
+
+  async getLinkValuesFromEntryNextLink(args: {
+    nextLink: string;
+    maxPageSize?: number;
+  }): Promise<ODataValueContextOfIListOfWEntryLinkInfo> {
+    let { nextLink, maxPageSize } = args;
+    return await getNextLinkListing<ODataValueContextOfIListOfWEntryLinkInfo>(
+      // @ts-ignore: allow sub class to use private variable from the super class
+      this.http,
+      this.processGetLinkValuesFromEntry,
+      nextLink,
+      maxPageSize
+    );
+  }
+
+  async getTagsAssignedToEntryNextLink(args: {
+    nextLink: string;
+    maxPageSize?: number;
+  }): Promise<ODataValueContextOfIListOfWTagInfo> {
+    let { nextLink, maxPageSize } = args;
+    return await getNextLinkListing<ODataValueContextOfIListOfWEntryLinkInfo>(
+      // @ts-ignore: allow sub class to use private variable from the super class
+      this.http,
+      this.processGetTagsAssignedToEntry,
+      nextLink,
+      maxPageSize
+    );
+  }
+}
+
+export interface IFieldDefinitionsClientEx extends IFieldDefinitionsClient {
+  GetFieldDefinitionsForEach(args: {
+    callback: (response: ODataValueContextOfIListOfWFieldInfo) => Promise<boolean>;
+    repoId: string;
+    prefer?: string;
+    culture?: string;
+    select?: string;
+    orderby?: string;
+    top?: number;
+    skip?: number;
+    count?: boolean;
+    maxPageSize?: number;
+  }): Promise<void>;
+  getFieldDefinitionsNextLink(args: {
+    nextLink: string;
+    maxPageSize?: number;
+  }): Promise<ODataValueContextOfIListOfEntry>;
+}
+
+export class FieldDefinitionClient extends FieldDefinitionsClient implements IFieldDefinitionsClientEx {
+  async GetFieldDefinitionsForEach(args: {
+    callback: (response: ODataValueContextOfIListOfWFieldInfo) => Promise<boolean>;
+    repoId: string;
+    prefer?: string;
+    culture?: string;
+    select?: string;
+    orderby?: string;
+    top?: number;
+    skip?: number;
+    count?: boolean;
+    maxPageSize?: number;
+  }): Promise<void> {
+    let { callback, repoId, prefer, culture, select, orderby, top, skip, count, maxPageSize } = args;
+    var response = await this.getFieldDefinitions({
+      repoId,
+      prefer: CreateMaxPageSizePreferHeaderPayload(maxPageSize),
+      culture,
+      select,
+      orderby,
+      top,
+      skip,
+      count,
+    });
+    let nextLink = response.odataNextLink;
+    while ((await callback(response)) && nextLink != null) {
+      response = await getNextLinkListing<ODataValueContextOfIListOfWFieldInfo>(
+        // @ts-ignore: allow sub class to use private variable from the super class
+        this.http,
+        this.processGetFieldDefinitions,
+        nextLink,
+        maxPageSize
+      );
+      nextLink = response.odataNextLink;
+    }
+  }
+  async getFieldDefinitionsNextLink(args: {
+    nextLink: string;
+    maxPageSize?: number;
+  }): Promise<ODataValueContextOfIListOfWFieldInfo> {
+    let { nextLink, maxPageSize } = args;
+    return await getNextLinkListing<ODataValueContextOfIListOfWFieldInfo>(
+      // @ts-ignore: allow sub class to use private variable from the super class
+      this.http,
+      this.processGetFieldDefinitions,
+      nextLink,
+      maxPageSize
+    );
+  }
+}
+
+export interface ISearchEx extends ISearchesClient {
+  GetSearchResultsForEach(args: {
+    callback: (response: ODataValueContextOfIListOfEntry) => Promise<boolean>;
+    repoId: string;
+    searchToken: string;
+    groupByEntryType?: boolean;
+    refresh?: boolean;
+    fields?: string[];
+    formatFields?: boolean;
+    prefer?: string;
+    culture?: string;
+    select?: string;
+    orderby?: string;
+    top?: number;
+    skip?: number;
+    count?: boolean;
+    maxPageSize?: number;
+  }): Promise<void>;
+  GetSearchContextHitsForEach(args: {
+    callback: (response: ODataValueContextOfIListOfContextHit) => Promise<boolean>;
+    repoId: string;
+    searchToken: string;
+    rowNumber: number;
+    prefer?: string;
+    select?: string;
+    orderby?: string;
+    top?: number;
+    skip?: number;
+    count?: boolean;
+    maxPageSize?: number;
+  }): Promise<void>;
+  GetSearchResultsNextLink(args: {
+    nextLink: string;
+    maxPageSize?: number;
+  }): Promise<ODataValueContextOfIListOfEntry>;
+  GetSearchContextHitsNextLink(args: {
+    nextLink: string;
+    maxPageSize?: number;
+  }): Promise<ODataValueContextOfIListOfContextHit>;
+}
+
+export class SearchClientEx extends SearchesClient implements ISearchEx {
+  async GetSearchResultsForEach(args: {
+    callback: (response: ODataValueContextOfIListOfEntry) => Promise<boolean>;
+    repoId: string;
+    searchToken: string;
+    groupByEntryType?: boolean;
+    refresh?: boolean;
+    fields?: string[];
+    formatFields?: boolean;
+    prefer?: string;
+    culture?: string;
+    select?: string;
+    orderby?: string;
+    top?: number;
+    skip?: number;
+    count?: boolean;
+    maxPageSize?: number;
+  }): Promise<void> {
+    let {
+      callback,
+      repoId,
+      searchToken,
+      groupByEntryType,
+      refresh,
+      fields,
+      formatFields,
+      prefer,
+      culture,
+      select,
+      orderby,
+      top,
+      skip,
+      count,
+      maxPageSize,
+    } = args;
+    var response = await this.getSearchResults({
+      repoId,
+      searchToken,
+      groupByEntryType,
+      refresh,
+      fields,
+      formatFields,
+      prefer: CreateMaxPageSizePreferHeaderPayload(maxPageSize),
+      culture,
+      select,
+      orderby,
+      top,
+      skip,
+      count,
+    });
+    let nextLink = response.odataNextLink;
+    while ((await callback(response)) && nextLink != null) {
+      response = await getNextLinkListing<ODataValueContextOfIListOfEntry>(
+        // @ts-ignore: allow sub class to use private variable from the super class
+        this.http,
+        this.processGetSearchResults,
+        nextLink,
+        maxPageSize
+      );
+      nextLink = response.odataNextLink;
+    }
+  }
+  async GetSearchContextHitsForEach(args: {
+    callback: (response: ODataValueContextOfIListOfContextHit) => Promise<boolean>;
+    repoId: string;
+    searchToken: string;
+    rowNumber: number;
+    prefer?: string;
+    select?: string;
+    orderby?: string;
+    top?: number;
+    skip?: number;
+    count?: boolean;
+    maxPageSize?: number;
+  }): Promise<void> {
+    let { callback, repoId, searchToken, rowNumber, prefer, select, orderby, top, skip, count, maxPageSize } = args;
+    var response = await this.getSearchContextHits({
+      repoId,
+      searchToken,
+      rowNumber,
+      prefer: CreateMaxPageSizePreferHeaderPayload(maxPageSize),
+      select,
+      orderby,
+      top,
+      skip,
+      count,
+    });
+    let nextLink = response.odataNextLink;
+    while ((await callback(response)) && nextLink != null) {
+      response = await getNextLinkListing<ODataValueContextOfIListOfContextHit>(
+        // @ts-ignore: allow sub class to use private variable from the super class
+        this.http,
+        this.processGetSearchContextHits,
+        nextLink,
+        maxPageSize
+      );
+      nextLink = response.odataNextLink;
+    }
+  }
+  async GetSearchResultsNextLink(args: {
+    nextLink: string;
+    maxPageSize?: number;
+  }): Promise<ODataValueContextOfIListOfEntry> {
+    let { nextLink, maxPageSize } = args;
+    return await getNextLinkListing<ODataValueContextOfIListOfEntry>(
+      // @ts-ignore: allow sub class to use private variable from the super class
+      this.http,
+      this.processGetSearchResults,
+      nextLink,
+      maxPageSize
+    );
+  }
+  async GetSearchContextHitsNextLink(args: {
+    nextLink: string;
+    maxPageSize?: number;
+  }): Promise<ODataValueContextOfIListOfContextHit> {
+    let { nextLink, maxPageSize } = args;
+    return await getNextLinkListing<ODataValueContextOfIListOfContextHit>(
+      // @ts-ignore: allow sub class to use private variable from the super class
+      this.http,
+      this.processGetSearchContextHits,
+      nextLink,
+      maxPageSize
+    );
+  }
+}
+
+export interface ITagDefinitionsEx extends ITagDefinitionsClient {
+  GetTagDefinitionsForEach(args: {
+    callback: (response: ODataValueContextOfIListOfWTagInfo) => Promise<boolean>;
+    repoId: string;
+    prefer?: string;
+    culture?: string;
+    select?: string;
+    orderby?: string;
+    top?: number;
+    skip?: number;
+    count?: boolean;
+    maxPageSize?: number;
+  }): Promise<void>;
+  getTagDefinitionsNextLink(args: {
+    nextLink: string;
+    maxPageSize?: number;
+  }): Promise<ODataValueContextOfIListOfWTagInfo>;
+}
+
+export class TagDefinitionsEx extends TagDefinitionsClient {
+  async GetTagDefinitionsForEach(args: {
+    callback: (response: ODataValueContextOfIListOfWTagInfo) => Promise<boolean>;
+    repoId: string;
+    prefer?: string;
+    culture?: string;
+    select?: string;
+    orderby?: string;
+    top?: number;
+    skip?: number;
+    count?: boolean;
+    maxPageSize?: number;
+  }): Promise<void> {
+    let { callback, repoId, prefer, culture, select, orderby, top, skip, count, maxPageSize } = args;
+    var response = await this.getTagDefinitions({
+      repoId,
+      prefer: CreateMaxPageSizePreferHeaderPayload(maxPageSize),
+      culture,
+      select,
+      orderby,
+      top,
+      skip,
+      count,
+    });
+    let nextLink = response.odataNextLink;
+    while ((await callback(response)) && nextLink != null) {
+      response = await getNextLinkListing<ODataValueContextOfIListOfWTagInfo>(
+        // @ts-ignore: allow sub class to use private variable from the super class
+        this.http,
+        this.processGetTagDefinitions,
+        nextLink,
+        maxPageSize
+      );
+      nextLink = response.odataNextLink;
+    }
+  }
+  async getTagDefinitionsNextLink(args: {
+    nextLink: string;
+    maxPageSize?: number;
+  }): Promise<ODataValueContextOfIListOfWTagInfo> {
+    let { nextLink, maxPageSize } = args;
+    return await getNextLinkListing<ODataValueContextOfIListOfWTagInfo>(
+      // @ts-ignore: allow sub class to use private variable from the super class
+      this.http,
+      this.processGetTagDefinitions,
+      nextLink,
+      maxPageSize
+    );
+  }
+}
+
+export interface ITemplateDefinitionsEx extends ITemplateDefinitionsClient {
+  GetTemplateDefinitionsForEach(args: {
+    callback: (response: ODataValueContextOfIListOfWTemplateInfo) => Promise<boolean>;
+    repoId: string;
+    templateName?: string;
+    prefer?: string;
+    culture?: string;
+    select?: string;
+    orderby?: string;
+    top?: number;
+    skip?: number;
+    count?: boolean;
+    maxPageSize?: number;
+  }): Promise<void>;
+  GetTemplateFieldDefinitionsForEach(args: {
+    callback: (response: ODataValueContextOfIListOfTemplateFieldInfo) => Promise<boolean>;
+    repoId: string;
+    templateId: number;
+    prefer?: string;
+    culture?: string;
+    select?: string;
+    orderby?: string;
+    top?: number;
+    skip?: number;
+    count?: boolean;
+    maxPageSize?: number;
+  }): Promise<void>;
+  GetTemplateFieldDefinitionsByTemplateNameForEach(args: {
+    callback: (response: ODataValueContextOfIListOfTemplateFieldInfo) => Promise<boolean>;
+    repoId: string;
+    templateName: string;
+    prefer?: string;
+    culture?: string;
+    select?: string;
+    orderby?: string;
+    top?: number;
+    skip?: number;
+    count?: boolean;
+    maxPageSize?: number;
+  }): Promise<void>;
+  getTemplateDefinitionsNextLink(args: {
+    nextLink: string;
+    maxPageSize?: number;
+  }): Promise<ODataValueContextOfIListOfWTemplateInfo>;
+  getTemplateFieldDefinitionsNextLink(args: {
+    nextLink: string;
+    maxPageSize?: number;
+  }): Promise<ODataValueContextOfIListOfTemplateFieldInfo>;
+  getTemplateFieldDefinitionsByTemplateNameNextLink(args: {
+    nextLink: string;
+    maxPageSize?: number;
+  }): Promise<ODataValueContextOfIListOfTemplateFieldInfo>;
+}
+
+export class TemplateDefinitionsEx extends TemplateDefinitionsClient {
+  async GetTemplateDefinitionsForEach(args: {
+    callback: (response: ODataValueContextOfIListOfWTemplateInfo) => Promise<boolean>;
+    repoId: string;
+    templateName?: string;
+    prefer?: string;
+    culture?: string;
+    select?: string;
+    orderby?: string;
+    top?: number;
+    skip?: number;
+    count?: boolean;
+    maxPageSize?: number;
+  }): Promise<void> {
+    let { callback, repoId, templateName, prefer, culture, select, orderby, top, skip, count, maxPageSize } = args;
+    var response = await this.getTemplateDefinitions({
+      repoId,
+      templateName,
+      prefer: CreateMaxPageSizePreferHeaderPayload(maxPageSize),
+      culture,
+      select,
+      orderby,
+      top,
+      skip,
+      count,
+    });
+    let nextLink = response.odataNextLink;
+    while ((await callback(response)) && nextLink != null) {
+      response = await getNextLinkListing<ODataValueContextOfIListOfWTemplateInfo>(
+        // @ts-ignore: allow sub class to use private variable from the super class
+        this.http,
+        this.processGetTemplateDefinitions,
+        nextLink,
+        maxPageSize
+      );
+      nextLink = response.odataNextLink;
+    }
+  }
+
+  async GetTemplateFieldDefinitionsForEach(args: {
+    callback: (response: ODataValueContextOfIListOfTemplateFieldInfo) => Promise<boolean>;
+    repoId: string;
+    templateId: number;
+    prefer?: string;
+    culture?: string;
+    select?: string;
+    orderby?: string;
+    top?: number;
+    skip?: number;
+    count?: boolean;
+    maxPageSize?: number;
+  }): Promise<void> {
+    let { callback, repoId, templateId, prefer, culture, select, orderby, top, skip, count, maxPageSize } = args;
+    var response = await this.getTemplateFieldDefinitions({
+      repoId,
+      templateId,
+      prefer: CreateMaxPageSizePreferHeaderPayload(maxPageSize),
+      culture,
+      select,
+      orderby,
+      top,
+      skip,
+      count,
+    });
+    let nextLink = response.odataNextLink;
+    while ((await callback(response)) && nextLink != null) {
+      response = await getNextLinkListing<ODataValueContextOfIListOfTemplateFieldInfo>(
+        // @ts-ignore: allow sub class to use private variable from the super class
+        this.http,
+        this.processGetTemplateFieldDefinitions,
+        nextLink,
+        maxPageSize
+      );
+      nextLink = response.odataNextLink;
+    }
+  }
+
+  async GetTemplateFieldDefinitionsByTemplateNameForEach(args: {
+    callback: (response: ODataValueContextOfIListOfTemplateFieldInfo) => Promise<boolean>;
+    repoId: string;
+    templateName: string;
+    prefer?: string;
+    culture?: string;
+    select?: string;
+    orderby?: string;
+    top?: number;
+    skip?: number;
+    count?: boolean;
+    maxPageSize?: number;
+  }): Promise<void> {
+    let { callback, repoId, templateName, prefer, culture, select, orderby, top, skip, count, maxPageSize } = args;
+    var response = await this.getTemplateFieldDefinitionsByTemplateName({
+      repoId,
+      templateName,
+      prefer: CreateMaxPageSizePreferHeaderPayload(maxPageSize),
+      culture,
+      select,
+      orderby,
+      top,
+      skip,
+      count,
+    });
+    let nextLink = response.odataNextLink;
+    while ((await callback(response)) && nextLink != null) {
+      response = await getNextLinkListing<ODataValueContextOfIListOfTemplateFieldInfo>(
+        // @ts-ignore: allow sub class to use private variable from the super class
+        this.http,
+        this.processGetTemplateFieldDefinitionsByTemplateName,
+        nextLink,
+        maxPageSize
+      );
+      nextLink = response.odataNextLink;
+    }
+  }
+  async getTemplateDefinitionsNextLink(args: {
+    nextLink: string;
+    maxPageSize?: number;
+  }): Promise<ODataValueContextOfIListOfWTemplateInfo> {
+    let { nextLink, maxPageSize } = args;
+    return await getNextLinkListing<ODataValueContextOfIListOfWTemplateInfo>(
+      // @ts-ignore: allow sub class to use private variable from the super class
+      this.http,
+      this.processGetTemplateDefinitions,
+      nextLink,
+      maxPageSize
+    );
+  }
+  async getTemplateFieldDefinitionsNextLink(args: {
+    nextLink: string;
+    maxPageSize?: number;
+  }): Promise<ODataValueContextOfIListOfTemplateFieldInfo> {
+    let { nextLink, maxPageSize } = args;
+    return await getNextLinkListing<ODataValueContextOfIListOfTemplateFieldInfo>(
+      // @ts-ignore: allow sub class to use private variable from the super class
+      this.http,
+      this.processGetTemplateFieldDefinitions,
+      nextLink,
+      maxPageSize
+    );
+  }
+  async getTemplateFieldDefinitionsByTemplateNameNextLink(args: {
+    nextLink: string;
+    maxPageSize?: number;
+  }): Promise<ODataValueContextOfIListOfTemplateFieldInfo> {
+    let { nextLink, maxPageSize } = args;
+    return await getNextLinkListing<ODataValueContextOfIListOfTemplateFieldInfo>(
+      // @ts-ignore: allow sub class to use private variable from the super class
+      this.http,
+      this.processGetTemplateFieldDefinitionsByTemplateName,
+      nextLink,
+      maxPageSize
+    );
+  }
 }
