+++ test/AccessTokens/AccessToken.test.ts
@@ -1,9 +1,8 @@
 import { repoId } from '../testHelper.js';
-import { IRepositoryApiClient } from '../../src/ClientBase.js';
-import { ODataValueOfBoolean, ODataValueOfDateTime } from '../../src/index.js';
+import { ODataValueOfBoolean, ODataValueOfDateTime, IRepositoryApiClient } from '../../src/index.js';
 import { createTestRepoApiClient } from '../BaseTest.js';
 
-describe('Access Token Integration Tests', () => {
+describe.skip('Access Token Integration Tests', () => {
   let _RepositoryApiClient: IRepositoryApiClient;
   _RepositoryApiClient = createTestRepoApiClient();
   test('Refresh Server Session', async () => {
