+++ test/Attributes/AttributeKeys.test.ts
@@ -1,9 +1,8 @@
 import { repoId } from '../testHelper.js';
-import { ODataValueContextOfListOfAttribute } from '../../src/index.js';
-import { IRepositoryApiClient } from '../../src/ClientBase.js';
+import { ODataValueContextOfListOfAttribute,IRepositoryApiClient } from '../../src/index.js';
 import { createTestRepoApiClient } from '../BaseTest.js';
 
-describe('Attribute Key Integration Tests', () => {
+describe.skip('Attribute Key Integration Tests', () => {
   let _RepositoryApiClient: IRepositoryApiClient;
   _RepositoryApiClient = createTestRepoApiClient();
   test('Get the attribute keys', async () => {
