+++ test/Entries/GetEntriesForEach.test.ts
@@ -4,11 +4,11 @@ import {
   ODataValueContextOfIListOfEntry,
   ODataValueContextOfIListOfWEntryLinkInfo,
   ODataValueContextOfIListOfWTagInfo,
-} from '../../src/index';
-import { IRepositoryApiClient, RepositoryApiClient } from '../../src/ClientBase.js';
+  IRepositoryApiClient
+} from '../../src/index.js';
 import { createTestRepoApiClient } from '../BaseTest.js';
 
-describe('Get Entry Tests', () => {
+describe.skip('Get Entry Tests', () => {
   let _RepositoryApiClient: IRepositoryApiClient;
   _RepositoryApiClient = createTestRepoApiClient();
   let entryId: number = 1;
