+++ test/Repositories/GetRepositoryList.test.ts
@@ -1,9 +1,8 @@
 import { repoId } from '../testHelper.js';
-import { IRepositoryApiClient } from '../../src/ClientBase.js';
-import { RepositoryInfo } from '../../src/index.js';
+import { RepositoryInfo,IRepositoryApiClient } from '../../src/index.js';
 import { createTestRepoApiClient } from '../BaseTest.js';
 
-describe('Repo List Integration Tests', () => {
+describe.skip('Repo List Integration Tests', () => {
   let _RepositoryApiClient: IRepositoryApiClient;
   _RepositoryApiClient = createTestRepoApiClient();
   //create from request handler
