+++ test/Searches/SearchTests.test.ts
@@ -1,15 +1,15 @@
 import { repoId } from '../testHelper.js';
-import { IRepositoryApiClient } from '../../src/ClientBase.js';
 import {
   AdvancedSearchRequest,
   ODataValueContextOfIListOfContextHit,
   ODataValueContextOfIListOfEntry,
+  IRepositoryApiClient
 } from '../../src/index.js';
-import { jest } from '@jest/globals';
+
 import { createTestRepoApiClient } from '../BaseTest.js';
 
 let searchToken = 'test';
-describe('Search Integration Tests', () => {
+describe.skip('Search Integration Tests', () => {
   let _RepositoryApiClient: IRepositoryApiClient;
   _RepositoryApiClient = createTestRepoApiClient();
